fix: remove data parameter from node data interface
