+++ packages/elements-utils/src/index.ts
@@ -1,3 +1,2 @@
 export { generateApiToC, generateProjectToC, generateTocSkeleton, resolveHttpServices } from './toc';
 export * from './toc/types';
-export * from './types';
