+++ packages/elements/src/hooks/useDereferencedHttpOperation.ts
@@ -1,7 +1,8 @@
-import { ParsedNode } from '@stoplight/elements-utils';
 import $RefParser from '@stoplight/json-schema-ref-parser';
 import { IHttpOperation, NodeType } from '@stoplight/types';
 import * as React from 'react';
+
+import { ParsedNode } from '../types';
 
 /**
  * Parses the branch node data and ONLY dereferences if it's an HTTP Operation
