+++ packages/elements/src/utils/oas/index.ts
@@ -1,12 +1,11 @@
 import { NodeData } from '@stoplight/elements-utils';
 import { Oas2HttpOperationTransformer, Oas3HttpOperationTransformer } from '@stoplight/http-spec/oas/types';
 import { encodePointerFragment, pointerToPath } from '@stoplight/json';
-import { IHttpOperation, IHttpService, NodeType } from '@stoplight/types';
+import { NodeType } from '@stoplight/types';
 import { get, isObject, last, map } from 'lodash';
 import { OpenAPIObject } from 'openapi3-ts';
 import { Spec } from 'swagger-schema-official';
 
-import { JSONSchema } from '../../types';
 import { ISourceNodeMap, NodeTypes } from './types';
 
 export const isOas2 = (parsed: unknown): parsed is Spec =>
@@ -104,7 +103,6 @@ export const computeNodeData = (uriMap: IUriMap, tags: string[] = []): NodeData[
           nodes.push({
             name: node['name'],
             type,
-            data: node as IHttpService,
             uri,
             tags,
           });
@@ -113,7 +111,6 @@ export const computeNodeData = (uriMap: IUriMap, tags: string[] = []): NodeData[
           nodes.push({
             name: node['summary'] || node['path'],
             type,
-            data: node as IHttpOperation,
             uri,
             tags: map(node['tags'], tag => tag['name']),
           });
@@ -122,7 +119,6 @@ export const computeNodeData = (uriMap: IUriMap, tags: string[] = []): NodeData[
           nodes.push({
             name: node['title'] || last(uri.split('/')) || '',
             type,
-            data: node as JSONSchema,
             uri,
             tags: node['x-tags'],
           });
